Jumping around a bit

+++ Task-Master.c
@@ -5,6 +5,9 @@
 #include <unistd.h>
 #include <signal.h>
 
+// File Type Controllers
+#include "cHandler.h"
+
 int main(int argc, char *argv[]) {
 	printf("Welcome to Task Master.\n");
 
@@ -14,27 +17,31 @@ int main(int argc, char *argv[]) {
 	char *program = argv[2];
 	// printf("Program being run: %s\n", program);
 
-	// Issue maybe arises in which program un-compiled. Work on maybe?
-	if ((strcmp(action, "start") != 1) && (program != NULL)) {
-		char command[100];
-		pid_t pid = fork();
-		if (pid < 0) {
-			printf("Fork Failed.\n");
-		} else if (pid == 0) {
-			if (strstr(program, "./") != NULL) {
-				system(program);
+	if (argv[1] != NULL && argv[2] != NULL) {
+		// Issue maybe arises in which program un-compiled. Work on maybe?
+		if ((strcmp(action, "start") != 1) && (program != NULL)) {
+			char command[100];
+			pid_t pid = fork();
+			if (pid < 0) {
+				printf("Fork Failed.\n");
+			} else if (pid == 0) {
+				if (strstr(program, "./") != NULL) {
+					system(program);
+				} else {
+					sprintf(command, "./%s", program);
+					system(command);
+				}
 			} else {
-				sprintf(command, "./%s", program);
-				system(command);
+				exit(0);
 			}
-		} else {
-			exit(0);
-		}
-	} else if(strcmp(action, "stop") == 0) {
-		printf("Self-destruct sequence initiated...\n");
-		system("sudo pkill test");
+		} else if(strcmp(action, "stop") == 0) {
+			printf("Self-destruct sequence initiated...\n");
+			system("sudo pkill test");
+		} 
+
 	} else {
-		printf("How did you veen get here?\n");
+		printf("How did you even get here?\n");
+		testFunction();
 	}
 
 	return 0;
